// frontend/src/api/generated/models/TimeSpan.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inundated
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TimeSpan
 */
export interface TimeSpan {
    /**
     * 
     * @type {string}
     * @memberof TimeSpan
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof TimeSpan
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof TimeSpan
     */
    startTime: Date;
    /**
     * 
     * @type {Date}
     * @memberof TimeSpan
     */
    endTime: Date;
    /**
     * 
     * @type {Set<string>}
     * @memberof TimeSpan
     */
    tagIds?: Set<string>;
}

/**
 * Check if a given object implements the TimeSpan interface.
 */
export function instanceOfTimeSpan(value: object): value is TimeSpan {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('startTime' in value) || value['startTime'] === undefined) return false;
    if (!('endTime' in value) || value['endTime'] === undefined) return false;
    return true;
}

export function TimeSpanFromJSON(json: any): TimeSpan {
    return TimeSpanFromJSONTyped(json, false);
}

export function TimeSpanFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimeSpan {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'startTime': (new Date(json['startTime'])),
        'endTime': (new Date(json['endTime'])),
        'tagIds': json['tagIds'] == null ? undefined : new Set(json['tagIds']),
    };
}

export function TimeSpanToJSON(json: any): TimeSpan {
    return TimeSpanToJSONTyped(json, false);
}

export function TimeSpanToJSONTyped(value?: TimeSpan | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'startTime': value['startTime'].toISOString(),
        'endTime': value['endTime'].toISOString(),
        'tagIds': value['tagIds'] == null ? undefined : Array.from(value['tagIds'] as Set<any>),
    };
}