// frontend/src/api/generated/models/CreateTimeSpan.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inundated
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateTimeSpan
 */
export interface CreateTimeSpan {
    /**
     * 
     * @type {string}
     * @memberof CreateTimeSpan
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof CreateTimeSpan
     */
    startTime: Date;
    /**
     * 
     * @type {Date}
     * @memberof CreateTimeSpan
     */
    endTime: Date;
    /**
     * 
     * @type {Set<string>}
     * @memberof CreateTimeSpan
     */
    tagIds?: Set<string>;
}

/**
 * Check if a given object implements the CreateTimeSpan interface.
 */
export function instanceOfCreateTimeSpan(value: object): value is CreateTimeSpan {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('startTime' in value) || value['startTime'] === undefined) return false;
    if (!('endTime' in value) || value['endTime'] === undefined) return false;
    return true;
}

export function CreateTimeSpanFromJSON(json: any): CreateTimeSpan {
    return CreateTimeSpanFromJSONTyped(json, false);
}

export function CreateTimeSpanFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateTimeSpan {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'startTime': (new Date(json['startTime'])),
        'endTime': (new Date(json['endTime'])),
        'tagIds': json['tagIds'] == null ? undefined : new Set(json['tagIds']),
    };
}

export function CreateTimeSpanToJSON(json: any): CreateTimeSpan {
    return CreateTimeSpanToJSONTyped(json, false);
}

export function CreateTimeSpanToJSONTyped(value?: CreateTimeSpan | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'startTime': value['startTime'].toISOString(),
        'endTime': value['endTime'].toISOString(),
        'tagIds': value['tagIds'] == null ? undefined : Array.from(value['tagIds'] as Set<any>),
    };
}